bob/harness/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7

Flask==2.2.3
waitress==2.1.2
importlib-metadata==4.12.0
zipp==1.0.0
more-itertools==8.10.0
requests==2.28.2
Jinja2==3.1.2
itsdangerous==2.1.2
Werkzeug==2.2.3
# click==8.1.3
click==8.1.7
MarkupSafe==2.1.2
urllib3==1.26.15
#certifi==2022.12.7

certifi==2024.2.2

#idna==3.4
idna==3.6



pipreqs==0.4.11
yarg==0.1.9
#charset-normalizer==3.1.0

charset-normalizer==3.3.2


docopt==0.6.2


flatbuffers==23.5.26
# gast==0.5.4
gast==0.4.0
setuptools==69.1.1
wheel==0.42.0
google-api-core==2.17.1
google-auth==2.28.1
# google-auth-oauthlib==1.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
#google-cloud-kms==2.21.2
google-cloud-kms==2.19.2
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
#Jinja2==3.1.3
Jinja2==3.1.2
# keras==2.15.0
keras==2.13.1
libclang==16.0.6
Markdown==3.5.2
# MarkupSafe==2.1.5
MarkupSafe==2.1.2
ml-dtypes==0.2.0
# numpy==1.26.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
# requests==2.31.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
# tensorboard==2.15.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
# tensorflow==2.15.0.post1
tensorflow==2.13.1
# tensorflow-estimator==2.15.0
tensorflow-estimator==2.13.0
# tensorflow-io-gcs-filesystem==0.36.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tink==1.9.0
tink[gcpkms]==1.9.0
# typing_extensions==4.10.0
typing_extensions==4.5.0
# urllib3==2.2.1
urllib3==1.26.15
# Werkzeug==3.0.1
# Werkzeug==2.2.2
Werkzeug==2.2.3
wrapt==1.14.1
